+++ client/python/test/test_client.py
@@ -324,6 +324,21 @@ class TestClientPredictions:
         with open(output) as f:
             assert f.read() == "Hello file!"
 
+    def test_upload_preserves_orig_name(self):
+        demo = gr.Interface(lambda x: x, "image", "text")
+        with connect(demo) as client:
+            test_file = str(Path(__file__).parent / "files" / "cheetah1.jpg")
+            output = client.endpoints[0]._upload_file({"path": test_file}, data_index=0)
+            assert output["orig_name"] == "cheetah1.jpg"
+
+            output = client.endpoints[0]._upload_file(
+                {
+                    "path": "https://raw.githubusercontent.com/gradio-app/gradio/main/test/test_files/bus.png"
+                },
+                data_index=0,
+            )
+            assert output["orig_name"] == "bus.png"
+
     @pytest.mark.flaky
     def test_cancel_from_client_queued(self, cancel_from_client_demo):
         with connect(cancel_from_client_demo) as client:
@@ -992,8 +1007,11 @@ class TestAPIInfo:
             "gradio-tests/not-actually-private-space",
         )
         assert len(client.endpoints) == 3
-        assert len([e for e in client.endpoints if e.is_valid]) == 2
-        assert len([e for e in client.endpoints if e.is_valid and e.api_name]) == 1
+        assert len([e for e in client.endpoints.values() if e.is_valid]) == 2
+        assert (
+            len([e for e in client.endpoints.values() if e.is_valid and e.api_name])
+            == 1
+        )
         assert client.view_api(return_format="dict") == {
             "named_endpoints": {
                 "/predict": {