+++ gradio/context.py
@@ -8,6 +8,7 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:  # Only import for type checking (is False at runtime).
     from gradio.blocks import BlockContext, Blocks, BlocksConfig
     from gradio.helpers import Progress
+    from gradio.renderable import Renderable
     from gradio.routes import Request
 
 
@@ -24,7 +25,7 @@ class LocalContext:
     blocks_config: ContextVar[BlocksConfig | None] = ContextVar(
         "blocks_config", default=None
     )
-    is_render: ContextVar[bool] = ContextVar("is_render", default=False)
+    renderable: ContextVar[Renderable | None] = ContextVar("renderable", default=None)
     render_block: ContextVar[BlockContext | None] = ContextVar(
         "render_block", default=None
     )
@@ -35,21 +36,21 @@ class LocalContext:
 
 
 def get_render_context() -> BlockContext | None:
-    if LocalContext.is_render.get():
+    if LocalContext.renderable.get():
         return LocalContext.render_block.get()
     else:
         return Context.block
 
 
 def set_render_context(block: BlockContext | None):
-    if LocalContext.is_render.get():
+    if LocalContext.renderable.get():
         LocalContext.render_block.set(block)
     else:
         Context.block = block
 
 
 def get_blocks_context() -> BlocksConfig | None:
-    if LocalContext.is_render.get():
+    if LocalContext.renderable.get():
         return LocalContext.blocks_config.get()
     elif Context.root_block:
         return Context.root_block.default_config